requirements.txt
Pillow
numpy
pyautogui
keyboards